services/blockchain/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
reportlab>=4.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
PyJWT>=2.8.0
redis>=5.0.0

